Move path enumeration methods in PDFIssues into a mixin for re-use by other ingest uses.
