+++ lib/newspaper_works/ingest.rb
@@ -9,6 +9,7 @@ require 'newspaper_works/ingest/publication_info'
 require 'newspaper_works/ingest/pub_finder'
 require 'newspaper_works/ingest/pdf_images'
 require 'newspaper_works/ingest/named_issue_metadata'
+require 'newspaper_works/ingest/path_enumeration'
 require 'newspaper_works/ingest/pdf_issue'
 require 'newspaper_works/ingest/pdf_issues'
 require 'newspaper_works/ingest/pdf_issue_ingester'
